chore: create extensions folder
